Motion State Function
Implemented the motion state functions. It still requires some testing in terms of the direction of Left and Right (they might be reversed order). Hopefully implemented it so that it can change between modes whenever.  Joystick is pretty much laid out as well. We need to play around with the delay (line 113). The template for DPAD is there, but the correct format is still to be determined.